+++ projects/ui/src/lib/components/button/index.ts
@@ -1,2 +1,3 @@
 export * from './kui-button-appearance.type';
+export * from './kui-button-shape.type';
 export * from './kui-button.directive';